+++ Starter Apps/REST-APIs/03 - DB Intro/code/create_tables.py
@@ -8,5 +8,11 @@ cursor = connection.cursor()
 create_tbl_query = "CREATE TABLE IF NOT EXISTS users (id INTEGER PRIMARY KEY, username text, password text)"
 cursor.execute(create_tbl_query)
 
+# create new table for items
+create_tbl_query = "CREATE TABLE IF NOT EXISTS items (name text, price real)"
+cursor.execute(create_tbl_query)
+
+cursor.execute("INSERT INTO items VALUES ('test', 10.99)")
+
 connection.commit()
 connection.close()
